Refactor SimulationPlotConfiguration to use generic PlotConfiguration
